Add Windows CI support and platform fixes
