+++ .github/workflows/ci.yml
@@ -7,12 +7,16 @@ on:
 permissions:
   contents: read
 
+defaults:
+  run:
+    shell: bash
+
 jobs:
   build-test:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v4
@@ -35,6 +39,7 @@ jobs:
         run: go build ./...
 
       - name: Validate plugin manifests
+        shell: bash
         run: |
           set -euo pipefail
           npm -s i -g ajv-cli@5.0.0 || true
@@ -69,6 +74,10 @@ jobs:
 
       - name: Test (race)
         run: go test -race ./...
+
+      - name: Windows plugin smoke test
+        if: matrix.os == 'windows-latest'
+        run: go test ./internal/e2e -run TestGlyphctlSmoke -count=1
 
       - name: Generate sample HTML report
         if: matrix.os == 'ubuntu-latest'
