+++ internal/plugins/runner/sysproc_unix.go
@@ -1,0 +1,12 @@
+//go:build !windows
+
+package runner
+
+import (
+	"os/exec"
+	"syscall"
+)
+
+func configureSysProc(cmd *exec.Cmd) {
+	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
+}
