+++ internal/plugins/runner/sysproc_windows.go
@@ -1,0 +1,10 @@
+//go:build windows
+
+package runner
+
+import "os/exec"
+
+func configureSysProc(cmd *exec.Cmd) {
+	// Windows does not support process groups in the same way as POSIX systems.
+	// The default configuration is sufficient for our sandbox.
+}
