+++ internal/plugins/runner/testutil_test.go
@@ -1,0 +1,14 @@
+package runner
+
+import (
+	"path/filepath"
+	"runtime"
+)
+
+func executablePath(dir, base string) string {
+	name := base
+	if runtime.GOOS == "windows" {
+		name += ".exe"
+	}
+	return filepath.Join(dir, name)
+}
