+++ plugins/cartographer/plugin.js.sig
@@ -1,1 +1,1 @@
-MEYCIQCZzUYhndKJdh89+4i7pWaYkLsLY/b91jhT3eSnTGCdVwIhAObuOCpt1FbtqGxR0cysoeAF9875q9Ul0q6sU61HjAew
+MEYCIQCPxhfxv3QP6x8YqiumjQOCgL37kCuTR54asvHzzBz8sQIhAJU0NwV0UVCfro+Ipk4uixLWss6mss316Nbp+0xxK6Hx
