+++ plugins/cryptographer/plugin.js.sig
@@ -1,1 +1,1 @@
-MEUCIQDDnkv12z9/Ys1+HYCPGrEEwrmMrFkufJUS4fLQSHUfIgIgBnb1xvwLZs2dDWn4MEtWs6ThHCb2LKVJyIsvLnre0K8=
+MEUCIFmRuWwDZSqFP/ocYBJp1Y/4ncvr2qUMXSiAjrQrvEfFAiEA0tYqWJ4bHn2+Nhu9UEGDpd7/uybprt+r5Me6Xiw430k=
