+++ plugins/example-hello/main.go.sig
@@ -1,1 +1,1 @@
-MEUCIQDAF82ehnt5n75sVWs6fcTulDW9tDW0CR64KCrHkR+xpwIgHEwF4nOUZ7dUUzsYSKWy/dzpXU4zxcYESJA9o1TPkbw=
+MEQCIEg4DCF/1JTNP61df2L7bN4Yq6Qx+IUMLOdry6g9L5FHAiAe77TwA6P32dt39kYZTxfU1l3NnJyy0Kez4JAgONempQ==
