+++ plugins/excavator/plugin.js.sig
@@ -1,1 +1,1 @@
-MEUCIQDJar8aDS3iFbP3Ccm+wUu0x4zZqd0yKMXnYCEbAnJcbwIgVenqTfEtEy71ePN1m4lefQWl6tae4n3Z39CT6JF1Dis=
+MEUCIQDfgdJQglW07cy7Zh6AbpLoN/6jVvqiTm5JnKtbEJzMXwIgS+v3ihfE4yhUsAQTA3IFjBjFtXG9eqgYFAJv+DOVU30=
