+++ plugins/galdr-proxy/plugin.js.sig
@@ -1,1 +1,1 @@
-MEUCIHtPdypHzUC+0tIAfQ7yAPB9+ldixfjU6QI9NpZQD4keAiEA8bry32FIDQBhptG8aKk4gTci58zas5AsM5gf/Tg6Nyk=
+MEQCIE//XvBeSwlHLRyiR93NzML3hSiy2oxVe2DvqDTgjpRNAiBCD9Jp5B3tWyeO6UDgbRhSh1p7WsRNE4CFLiSNhntYpg==
