+++ plugins/grapher/main.go.sig
@@ -1,1 +1,1 @@
-MEUCIQDMKdqhCFxkkF0bQkCKNNDNLaS3B5+/eXC7I0JlmBUSJAIgOQJzbHlZx/OfTOOX4X8Wu1ocdV3uBU/J65t1RssZOTI=
+MEYCIQCPf6753SSZvAQ2fjMzrBGn9R0dKadS6kAuhbI9byQynQIhAL8OCbYQWVmonDYxqv0FEMYuSMxvEwig35g4C5lfTEq+
