+++ plugins/keys/glyph-plugin.pub
@@ -1,4 +1,4 @@
 -----BEGIN PUBLIC KEY-----
-MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAESsQZm43tpsSkOF4JXVtkCT8w9s/F
-Ic9tuVjLXi7rXrME5YGd4Puqw5xZEQukFRwrzjBlIqDOougIOzRQjJQffw==
+MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEDAcD/i7miF6EQktlZi1zsv6b1CBG
+6/omRgq4rhD9OnsDlmPqq9zRv/yY8pVuGgg8rY4dp7Gl1KEd7NGhBxBKBQ==
 -----END PUBLIC KEY-----
