+++ plugins/osint-well/plugin.js.sig
@@ -1,1 +1,1 @@
-MEUCIQCln2g3T0IhbFWf8uHXC6vJ5sIK4C1qNQG19rHRCLuIdgIgNvqG1faWvtFscFdhLemub+jljJIKDIhbTQeaXTI3/3A=
+MEUCIAx4sCn+oQtH9+c5Fl2BrpNYIMClryh3KuSsWtMAZ2hZAiEAiTiHxTVhqWeSgEqAb5n2cNIDkQQ/1nbK4zFkhLPfRxE=
