+++ plugins/raider/plugin.js.sig
@@ -1,1 +1,1 @@
-MEQCIAyl/RQ2Iaz1QoGlnIIMYvod0ih7rbNRKlxev92LY8rgAiBcZ+JAI48l7tn9e8nbKSug1iv4CA8Ato0N9QX145mPKQ==
+MEUCIADda137v0Cm86Rtbh3gikiX4IGXN7uwDmc8GgQMsGUZAiEA2qEa9GC2+3Tzl3ttx6DoVjseN7NeNj5+h4y+QuC9Yw8=
