+++ plugins/ranker/plugin.js.sig
@@ -1,1 +1,1 @@
-MEQCICruRNhAxZh5saEWQTU7Ne4BOMh74Dd0iViNdhOdLolVAiAtld5scEVi6Iq9iyFbrB3aMe78QFj8V77BRXnb9fRkOA==
+MEUCIF/qDDn+n+eqCNMJ9dcqn9QmkvCxefl1OWD06yVN2UZHAiEA2n6ryrgEP9Tc9ljKgU+p4HUymZX+Rqga8BRIet8hnLg=
