+++ plugins/samples/emit-on-start/main.go.sig
@@ -1,1 +1,1 @@
-MEYCIQDt08XXeP1moOomkDnLyQ8t21L4PO7T/qGxkfjZp3WwfgIhAKaq9kvcZ0QQqEzEIvhvy1eBW2O5hJ9XM42tC/3Mb+gi
+MEUCIBeytDcbS4baWHgva2113R1e8CCJ0KVnnT7Yc6pCmUUoAiEAvWzD2UsqVPLUe8yocV7BLTbY6ty6wRhXaLGR5E0/b80=
