+++ plugins/samples/emit-on-start/signals_windows.go
@@ -1,0 +1,9 @@
+//go:build windows
+
+package main
+
+import "os"
+
+func terminationSignals() []os.Signal {
+	return []os.Signal{os.Interrupt}
+}
