+++ plugins/samples/passive-header-scan/main.go.sig
@@ -1,1 +1,1 @@
-MEUCIEPvr6QFvHyk73Dztdg9cyyEXPLoSReeTMDerUUtNVCPAiEAyAyeWFM2vs5jkXHRmhFRhpIctO7C5GG0BQ380kkx9i8=
+MEYCIQCGds0xiJOvrEHlATBH7lwSg9NHer3T/7c5h9y13n2h1wIhAMF8re3B+rQpyKqWjkO3P2OxGwy4Tmu2uYhjWKl5CTUt
