+++ plugins/scribe/plugin.js.sig
@@ -1,1 +1,1 @@
-MEQCIDiTigBEP7QcrG6GQhR5v011c/7o3bntxs1aN24fFDytAiAOD8zoeYtS7ykpL2UKjwnGVydpSv0EgvaqwM1cPX7ujg==
+MEUCIQCtUyl1oVTDV6YyPrr82VPhcHyhNEn/Zn7FupBEKoJ9bAIgUYM8UlPb/32dWmGm+e4LYPTDWKnzpE9/Z902gV0CcvI=
