+++ plugins/seer/main.go.sig
@@ -1,1 +1,1 @@
-MEUCIQC3BsZxWrJIxSbCgWFHX771UnsKHgtH/BTJQZDZjAGE2AIgOYdDCSF7866pvHiFM+datgu8fj1vEUgLu24i4kpHXr8=
+MEYCIQDqZxwR7HMSNvnkmcn1pzHNtzvB6b38vTuageJhOBECKQIhAPZHIp860hYkG/uHrroI8cxrB8cH2Fc/QHEj4Gp1fYzL
